+++ src/remarx/sentence/__init__.py
@@ -1,7 +1,8 @@
 """
-This module contains libraries for sentence segmentation and sentence corpus construction.
+This module contains libraries for sentence segmentation, sentence corpus construction,
+and sentence embedding generation.
 """
 
 from remarx.sentence import corpus, segment
 
-__all__ = ["corpus", "segment"]
+__all__ = ["corpus", "embeddings", "segment"]